fix: remove warnings

+++ lua/resurrect/file.lua
@@ -12,6 +12,9 @@ end
 function M.load_buffers()
   local buffers = {}
   local file = open_file('r')
+  if file == nil then
+    return
+  end
   for c in file:lines() do
     table.insert(buffers, c)
   end
@@ -20,12 +23,18 @@ end
 
 function M.del(path)
   local file = open_file('r')
+  if file == nil then
+    return
+  end
   local ct = file:read('a')
   file:close()
 
   os.remove('./.resurrect')
 
   file = open_file('w')
+  if file == nil then
+    return
+  end
   for c in ct:gmatch('[^\r\n]+') do
     if c ~= path then
       file:write(c .. '\n')
@@ -37,6 +46,9 @@ end
 
 function M.add(path)
   local file = open_file('a')
+  if file == nil then
+    return
+  end
   file:write(path .. '\n')
   file:flush()
   file:close()
